pyproject.toml
[project]
name = "trade-strats"
version = "0.1.0"
description = "Deterministic TheStrat 15m trading bot for US equities via Alpaca"
readme = "docs/prd/draft.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [{ name = "dkalfa" }]

dependencies = [
    "alpaca-py>=0.30",
    "pandas>=2.2",
    "numpy>=1.26",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "aiosqlite>=0.20",
    "structlog>=24.1",
    "rich>=13.7",
    "typer>=0.12",
    "httpx>=0.27",
    "pytz>=2024.1",
]

[project.scripts]
trade-strats = "trade_strats.cli:app"

[tool.uv]
dev-dependencies = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "hypothesis>=6.100",
    "time-machine>=2.14",
    "ruff>=0.5",
    "pyright>=1.1.370",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trade_strats"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",   # pycodestyle + pyflakes
    "I",             # isort
    "B",             # flake8-bugbear
    "UP",            # pyupgrade
    "SIM",           # flake8-simplify
    "RUF",           # ruff-specific
    "N",             # pep8-naming
    "C4",            # comprehensions
    "PTH",           # pathlib
    "TID",           # tidy-imports
]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.lint.per-file-ignores]
# Typer uses `= typer.Option(...)` as the standard CLI arg pattern.
"src/trade_strats/cli.py" = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
include = ["src", "tests"]
pythonVersion = "3.12"
typeCheckingMode = "strict"
reportMissingTypeStubs = false

[[tool.pyright.executionEnvironments]]
root = "tests"
reportUnknownMemberType = "none"
reportUnknownVariableType = "none"
reportUnknownArgumentType = "none"
reportPrivateUsage = "none"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"
